feat: wishlistXstock 관련 api 생성 (#9)
* chore: upsert_stock.py 경로 이동

* chore: stock data not fount 에러 추가

* feat: wishlistXstock 모델, 스키마 디자인

* chore: _check_wishlist_exist_and_access_permission 함수 위치 변경

* feat: wishlistXstock 생성 기능 구현

* chore: delete return_rate field

* chore: fix check_wishlist_exist_and_access_permission func name

* feat: add fetch wishlistXstock api

* feat: add update wishlistXstock update api

* refactor: order update func 모듈화

* refactor: change_wishlist_order 함수에 refactoring util 적용

* feat: add wishlistXstock update, delete, reorder api

* refactor: reorder_for_change_order func for utilizing

* chore: rename _reorder_wishlistXstock_for_delete func

* chore: change dir of resource folder

* chore: change dir of upsert_stock.py file and db

* chore: edit endpoint url for restful method

* fix: delete wishlistXstock in endpoints and merged with wishlist

* fix: rename func of queries.wishlistXstock.py

* fix: rename valiate_accessible_wishlist and change dir to queries.wishlist.py

* fix: 서비스단의 wishlistXstock 전부 WishStock 으로 명칭 변경

* refactor: rename wishlistXstock to wishstock

* fix: edit _reorder_wishstock_for_delete error

* chore: rename private method of utils

* build: add scripts about upsert_stock.py

* chore: delete wishStock schemas

* chore: edit WishStockCreate schemas

* refactor: edit WishStockResponse about stock and wishlist value

* refactor: edit wishstock api for restful method

* refactor: edit reorder method for util

+++ onboarding_app/queries/wishlist.py
@@ -1,15 +1,14 @@
 from datetime import datetime
 
+from fastapi.encoders import jsonable_encoder
 from sqlalchemy import text
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Query, Session
 
-from onboarding_app import exceptions, models, schemas
+from onboarding_app import exceptions, models, schemas, utils
 
 
-def _check_wishlist_exist_and_access_permission(
-    wishlist_query_res: Query, current_user: schemas.User
-):
+def validate_accessible_wishlist(wishlist_query_res: Query, current_user: schemas.User):
     if not wishlist_query_res.first():
         raise exceptions.DataDoesNotExistError
     elif wishlist_query_res.first().user_id != current_user.id:
@@ -84,7 +83,7 @@ def update_wishlist(
     wishlist_query_res = db.query(models.Wishlist).filter(
         models.Wishlist.id == wishlist_id
     )
-    _check_wishlist_exist_and_access_permission(wishlist_query_res, current_user)
+    validate_accessible_wishlist(wishlist_query_res, current_user)
     try:
         wishlist_query_res.update(wishlist.dict(exclude_unset=True))
         wishlist_query_res.first().updated_at = datetime.utcnow()
@@ -101,7 +100,7 @@ def delete_wishlist(
     wishlist_query_res = db.query(models.Wishlist).filter(
         models.Wishlist.id == wishlist_id
     )
-    _check_wishlist_exist_and_access_permission(wishlist_query_res, current_user)
+    validate_accessible_wishlist(wishlist_query_res, current_user)
     wishlist_query_res.delete()
     db.commit()
     _reorder_wishlist_order_num(db, current_user.id)
@@ -128,46 +127,240 @@ def change_wishlist_order(
     wishlist_query_res = db.query(models.Wishlist).filter(
         models.Wishlist.id == wishlist_id
     )
-    users_wishlist_query_res = db.query(models.Wishlist).filter(
+    wishlistList_query_res_by_foreign_key = db.query(models.Wishlist).filter(
         models.Wishlist.user_id == current_user.id
     )
 
-    _check_wishlist_exist_and_access_permission(wishlist_query_res, current_user)
-    if hope_order < 0 or hope_order >= users_wishlist_query_res.count():
-        raise exceptions.InvalidQueryError
+    validate_accessible_wishlist(wishlist_query_res, current_user)
 
-    origin_order = wishlist_query_res.first().order_num
-    if hope_order > origin_order:
-        _update_upper_num_order(db, current_user, origin_order, hope_order)
+    utils.reorder(
+        target_model_obj=wishlist_query_res.first(),
+        hope_order=hope_order,
+        modelList_query_res_by_foreign_key=wishlistList_query_res_by_foreign_key,
+    )
 
-    elif hope_order < origin_order:
-        _update_lower_num_order(db, current_user, origin_order, hope_order)
-    wishlist_query_res.first().order_num = hope_order
     db.commit()
     return wishlist_query_res.first()
 
 
-def _update_upper_num_order(
-    db: Session, current_user: schemas.User, origin_order: int, hope_order: int
-):
-    db.query(models.Wishlist).filter(
-        models.Wishlist.order_num > origin_order,
-        models.Wishlist.order_num <= hope_order,
-        models.Wishlist.user_id == current_user.id,
-    ).update(
-        {models.Wishlist.order_num: models.Wishlist.order_num - 1},
-        synchronize_session=False,
+def _get_wishstock_response(
+    db_wishstock: schemas.WishStock,
+    db_stock: schemas.Stock,
+) -> schemas.WishStockResponse:
+
+    return_rate = round(
+        (db_stock.price - db_wishstock.purchase_price)
+        / db_wishstock.purchase_price
+        * 100,
+        2,
+    )
+
+    return schemas.WishStockResponse(
+        stock=jsonable_encoder(db_stock),
+        order_num=db_wishstock.order_num,
+        purchase_price=db_wishstock.purchase_price,
+        holding_num=db_wishstock.holding_num,
+        return_rate=return_rate,
     )
 
 
-def _update_lower_num_order(
-    db: Session, current_user: schemas.User, origin_order: int, hope_order: int
-):
-    db.query(models.Wishlist).filter(
-        models.Wishlist.order_num < origin_order,
-        models.Wishlist.order_num >= hope_order,
-        models.Wishlist.user_id == current_user.id,
-    ).update(
-        {models.Wishlist.order_num: models.Wishlist.order_num + 1},
-        synchronize_session=False,
+def add_stock_to_wishlist(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    wishstock: schemas.WishStockCreate,
+) -> schemas.WishStockResponse:
+
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
     )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    db_stock = (
+        db.query(models.Stock).filter(models.Stock.id == wishstock.stock_id).first()
+    )
+    if not db_stock:
+        raise exceptions.StockNotFoundError
+
+    count_for_order = (
+        db.query(models.WishlistXstock)
+        .filter(models.WishlistXstock.wishlist_id == wishlist_id)
+        .count()
+    )
+
+    try:
+        created_wishstock = models.WishlistXstock(
+            wishlist_id=wishlist_id,
+            stock_id=db_stock.id,
+            purchase_price=wishstock.purchase_price,
+            holding_num=wishstock.holding_num,
+            order_num=count_for_order,
+        )
+        db.add(created_wishstock)
+        db.commit()
+        db.refresh(created_wishstock)
+    except ZeroDivisionError:
+        raise exceptions.InvalidQueryError
+    except IntegrityError:
+        raise exceptions.DuplicatedError
+
+    return _get_wishstock_response(created_wishstock, db_stock)
+
+
+def fetch_stock_in_wishlist(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+) -> list[schemas.WishStockResponse]:
+
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
+    )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    wishstocks_query_res = (
+        db.query(models.WishlistXstock)
+        .filter(models.WishlistXstock.wishlist_id == wishlist_id)
+        .order_by(models.WishlistXstock.order_num)
+        .all()
+    )
+
+    wishstocks = []
+    for db_wishstock in wishstocks_query_res:
+        db_stock = (
+            db.query(models.Stock)
+            .filter(models.Stock.id == db_wishstock.stock_id)
+            .first()
+        )
+
+        wishstocks.append(_get_wishstock_response(db_wishstock, db_stock))
+
+    return wishstocks
+
+
+def get_stock_in_wishlist(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    stock_id: int,
+) -> schemas.WishStockResponse:
+
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
+    )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    wishstock_query_res = db.query(models.WishlistXstock).filter(
+        models.WishlistXstock.wishlist_id == wishlist_id,
+        models.WishlistXstock.stock_id == stock_id,
+    )
+
+    if not wishstock_query_res.first():
+        raise exceptions.DataDoesNotExistError
+
+    db_wishstock = wishstock_query_res.first()
+    db_stock = db.query(models.Stock).filter(models.Stock.id == stock_id).first()
+
+    return _get_wishstock_response(db_wishstock, db_stock)
+
+
+def update_stock_in_wishlist(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    stock_id: int,
+    wishstock: schemas.WishStockUpdate,
+) -> schemas.WishStockResponse:
+
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
+    )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    wishstock_query_res = db.query(models.WishlistXstock).filter(
+        models.WishlistXstock.wishlist_id == wishlist_id,
+        models.WishlistXstock.stock_id == stock_id,
+    )
+
+    if not wishstock_query_res.first():
+        raise exceptions.DataDoesNotExistError
+
+    wishstock_query_res.update(wishstock.dict(exclude_unset=True))
+    db.commit()
+    db.refresh(wishstock_query_res.first())
+
+    db_wishstock = wishstock_query_res.first()
+    db_stock = db.query(models.Stock).filter(models.Stock.id == stock_id).first()
+
+    return _get_wishstock_response(db_wishstock, db_stock)
+
+
+def delete_stock_in_wishlist(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    stock_id: int,
+) -> None:
+
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
+    )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    wishstock_query_res = db.query(models.WishlistXstock).filter(
+        models.WishlistXstock.wishlist_id == wishlist_id,
+        models.WishlistXstock.stock_id == stock_id,
+    )
+
+    if not wishstock_query_res.first():
+        raise exceptions.DataDoesNotExistError
+
+    wishstock_query_res.delete()
+    db.commit()
+    _reorder_wishstock_for_delete(db, wishlist_id)
+
+    return None
+
+
+def _reorder_wishstock_for_delete(db: Session, wishlist_id: int):
+    users_wishlist_query_res = (
+        db.query(models.WishlistXstock)
+        .filter(models.WishlistXstock.wishlist_id == wishlist_id)
+        .order_by(models.WishlistXstock.order_num)
+    )
+    for i, wishstock in enumerate(users_wishlist_query_res):
+        wishstock.order_num = i
+    db.commit()
+
+
+def change_stock_order(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    stock_id: int,
+    hope_order: int,
+) -> schemas.WishStockResponse:
+    wishlist_query_res = db.query(models.Wishlist).filter(
+        models.Wishlist.id == wishlist_id
+    )
+    validate_accessible_wishlist(wishlist_query_res, current_user)
+
+    wishstockList_query_res_by_foreign_key = db.query(models.WishlistXstock).filter(
+        models.WishlistXstock.wishlist_id == wishlist_id
+    )
+
+    wishstock_query_res = wishstockList_query_res_by_foreign_key.filter(
+        models.WishlistXstock.stock_id == stock_id
+    )
+
+    utils.reorder(
+        target_model_obj=wishstock_query_res.first(),
+        hope_order=hope_order,
+        modelList_query_res_by_foreign_key=wishstockList_query_res_by_foreign_key,
+    )
+    db.commit()
+
+    db_wishstock = wishstock_query_res.first()
+    db_stock = db.query(models.Stock).filter(models.Stock.id == stock_id).first()
+
+    return _get_wishstock_response(db_wishstock, db_stock)
